requirements.txt
asgiref==3.11.0
astroid==3.3.11
attrs==25.4.0
autopep8==2.3.2
dill==0.4.0
dj-database-url==3.1.2
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
inflection==0.5.1
isort==6.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mccabe==0.7.0
platformdirs==4.5.1
pycodestyle==2.14.0
pylint==3.3.9
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.4
tomlkit==0.13.3
typing_extensions==4.15.0
uritemplate==4.2.0
gunicorn
whitenoise
psycopg2-binary==2.9.10
